ISPN-6297 Moved ConcurrentHashSet to commons
