+++ core/src/main/java/org/infinispan/util/concurrent/SynchronizedRestarter.java
@@ -1,6 +1,7 @@
 package org.infinispan.util.concurrent;
 
 import org.infinispan.commons.api.Lifecycle;
+import org.infinispan.commons.util.concurrent.ConcurrentHashSet;
 
 import java.util.concurrent.atomic.AtomicBoolean;
 import java.util.concurrent.locks.LockSupport;
